fix(dev): implement ethereum dev node in typescript
We implement the script to start the ethereum dev node in typescript
with the following improvements
* We wait for the dev node to be ready before deploying the contracts
* We equip the dev account with funds only once
* We eliminate the flaky process management from the shell script
* We provide a better error message when the dev account address cannot
  be read
